Adds requested changes
